1月5号早上的。

+++ src/main/java/com/my/util/TokenInfo.java
@@ -1,16 +1,18 @@
 package com.my.util;
 
 public class TokenInfo {
-	private int userId; // 用户唯一标识ID
+	private Long userId; // 用户唯一标识ID
 	private String username; // 用户登录名
 	private String ssoClient; // 来自登录请求的某应用系统标识
 	private String globalId; // 本次登录成功的全局会话sessionId
 
-	public int getUserId() {
+	
+
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(int userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
